Update slot manager

+++ app/services/slot_manager.rb
@@ -39,10 +39,6 @@ class SlotManager
     @tutor.slots.each do |slot|
       @slot_start_DOW_time = slot.start_time.strftime('%a %T')
       @slot_duration = slot.duration
-      puts @slot_start_DOW_time
-      puts @original_start_DOW_time
-      puts @slot_start_DOW_time == @original_start_DOW_time
-      puts @slot_duration == @original_duration.to_i
       if @slot_start_DOW_time == @original_start_DOW_time && @slot_duration == @original_duration.to_i
         @slots << slot
       end
@@ -68,7 +64,11 @@ class SlotManager
     @slot_ids = []
     @slots.each do |slot|
       @slot_ids << slot
-      slot.destroy
+      if slot.appointments.empty? || slot.appointments.where(status: 0).empty?
+        slot.destroy
+      else
+        slot.update(status: 'Zombie')
+      end
     end
     @slot_ids
   end
